06-17-2024 change float to np.float64

+++ cytominer_eval/operations/replicate_reproducibility.py
@@ -1,5 +1,5 @@
 """Functions to calculate replicate reproducibility."""
-
+import numpy as np
 import pandas as pd
 from typing import List
 
@@ -10,9 +10,9 @@ from cytominer_eval.utils.transform_utils import assert_melt
 def replicate_reproducibility(
     similarity_melted_df: pd.DataFrame,
     replicate_groups: List[str],
-    quantile_over_null: float = 0.95,
+    quantile_over_null: np.float64 = 0.95,
     return_median_correlations: bool = False,
-) -> float:
+) -> np.float64:
     r"""Summarize pairwise replicate correlations
 
     For a given pairwise similarity matrix, replicate information, and specific options,
